# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/staff/seller-applications": {
      "get": {
        "operationId": "StaffSellerApplicationController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-seller-applications"
        ]
      }
    },
    "/staff/seller-applications/{id}": {
      "get": {
        "operationId": "StaffSellerApplicationController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerApplicationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-seller-applications"
        ]
      }
    },
    "/staff/seller-applications/{id}/approve": {
      "post": {
        "operationId": "StaffSellerApplicationController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerApplicationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-seller-applications"
        ]
      }
    },
    "/staff/seller-applications/{id}/reject": {
      "post": {
        "operationId": "StaffSellerApplicationController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectSellerApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerApplicationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-seller-applications"
        ]
      }
    },
    "/staff/sellers/{userId}/suspend": {
      "post": {
        "operationId": "StaffSellerController_suspend",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendSellerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-sellers"
        ]
      }
    },
    "/staff/sellers/{userId}/reinstate": {
      "post": {
        "operationId": "StaffSellerController_reinstate",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-sellers"
        ]
      }
    },
...
    "/me/payments/{id}": {
      "get": {
        "operationId": "MePaymentController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me-payments"
        ]
      }
    },
    "/staff/payments": {
      "get": {
        "operationId": "StaffPaymentController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "SUCCEEDED",
                "FAILED",
                "CANCELLED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "contractId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payerUserId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-payments"
        ]
      }
    },
    "/staff/payments/{id}": {
      "get": {
        "operationId": "StaffPaymentController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPaymentResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-payments"
        ]
      }
    },
...
    "/seller/payouts": {
      "post": {
        "operationId": "SellerPayoutController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-payouts"
        ]
      },
      "get": {
        "operationId": "SellerPayoutController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "SUCCEEDED",
                "FAILED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-payouts"
        ]
      }
    },
    "/seller/payouts/{id}": {
      "get": {
        "operationId": "SellerPayoutController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "seller-payouts"
        ]
      }
    },
...
    "/staff/payouts/{id}": {
      "get": {
        "operationId": "StaffPayoutController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffPayoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "staff-payouts"
        ]
      }
    }
  },
  "info": {
    "title": "Bobo&Doda API",
    "description": "Markaziy Osiyo freelance marketplace — milestone escrow, double-entry ledger. Frontend `lib/api` chegarasini qoplaydi.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SellerApplicationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "legalName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reviewedAt": {
            "type": "object",
            "nullable": true
          },
          "rejectionReason": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "legalName",
          "displayName",
          "status",
          "submittedAt",
          "createdAt"
        ]
      },
...
      "RejectSellerApplicationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
      "SuspendSellerDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5,
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "PaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contractId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "failureReason": {
            "type": "object"
          },
          "succeededAt": {
            "type": "object"
          },
          "failedAt": {
            "type": "object"
          },
          "cancelledAt": {
            "type": "object"
          },
          "expiresAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "contractId",
          "provider",
          "status",
          "amount",
          "currency",
          "createdAt",
          "updatedAt"
        ]
      },
      "StaffPaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contractId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "failureReason": {
            "type": "object"
          },
          "succeededAt": {
            "type": "object"
          },
          "failedAt": {
            "type": "object"
          },
          "cancelledAt": {
            "type": "object"
          },
          "expiresAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "payerUserId": {
            "type": "string"
          },
          "providerPaymentId": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "contractId",
          "provider",
          "status",
          "amount",
          "currency",
          "createdAt",
          "updatedAt",
          "payerUserId"
        ]
      },
...
      "CreatePayoutDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Butun so‘m",
            "minimum": 1
          },
          "destinationReference": {
            "type": "string",
            "minLength": 3,
            "maxLength": 200
          }
        },
        "required": [
          "amount",
          "destinationReference"
        ]
      },
      "PayoutResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "destinationReference": {
            "type": "string"
          },
          "failureReason": {
            "type": "object"
          },
          "succeededAt": {
            "type": "object"
          },
          "failedAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "amount",
          "currency",
          "destinationReference",
          "createdAt",
          "updatedAt"
        ]
      },
      "StaffPayoutResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Butun so‘m"
          },
          "currency": {
            "type": "string"
          },
          "destinationReference": {
            "type": "string"
          },
          "failureReason": {
            "type": "object"
          },
          "succeededAt": {
            "type": "object"
          },
          "failedAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sellerId": {
            "type": "string"
          },
          "providerPayoutId": {
            "type": "object"
          },
          "provider": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "amount",
          "currency",
          "destinationReference",
          "createdAt",
          "updatedAt",
          "sellerId",
          "provider"
        ]
      }
    }
  }
}